GUI loads, separated functions although they don't work
